feat(api): implementa controller e use case para transacoes

+++ NTT_DATA-Backend_Java_com_Spring_AI/projetos/Modulo_04/05-Desafio-Desenvolvendo.sua.API.Inteligente.com.Recrutamento.de.IA.com.Spring.Boot/budgeting/src/main/java/dio/budgeting/domain/Transaction.java
@@ -8,10 +8,10 @@ import lombok.Getter;
 public class Transaction {
     private TransactionId id;
     private String description;
-    private long amount;
+    private double amount;
     private Category category;
 
-    public Transaction(String description, long amount, Category category) {
+    public Transaction(String description, double amount, Category category) {
         this.id = new TransactionId();
         this.description = description;
         this.amount = amount;
